Use stream encoding instead of locale preferred encoding
