+++ test_all.sh
@@ -60,10 +60,7 @@ for encoding in ascii "UTF-8"; do
 			[[ $color == "1" ]] && color_filename="color" || color_filename="nocolor"
 			filename="test/output/python3-${term}-${encoding}-${color_filename}.out"
 
-			export LANG="en_US.${encoding}"
-			export LC_ALL="${LANG}"
-			export PYTHONCOERCECLOCALE=0
-			export PYTHONUTF8=0
+			export PYTHONIOENCODING="${encoding}"
 			export TERM="${term}"
 			export FORCE_COLOR="${color}"
